fixed partition prune

+++ src/partition/disk_partition.h
@@ -11,4 +11,4 @@ void diskANN_partitions_with_ram_budget(const std::string data_file, const doubl
 
 template <typename T>
 void scaleGANN_partitions_with_ram_budget(const std::string data_file, const double sampling_rate, double ram_budget,
-    size_t graph_degree, const std::string prefix_path, size_t k_base, uint32_t num_parts = 0, uint32_t epsilon = 2, size_t max_k_means_reps = 15);
+    size_t graph_degree, const std::string prefix_path, size_t k_base, uint32_t num_parts = 0, float epsilon = 2, size_t max_k_means_reps = 15);
